refactor(runner): extract firecracker config builder to separate module
Move FirecrackerConfig interface and config building logic from vm.ts
to a new config.ts module. This allows reuse by the snapshot command
without duplicating code.
